[refactor] move TempGPGWrapper to it's own file
